requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
avro==1.11.0
blinker==1.7.0
boto3==1.34.7
botocore==1.34.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.1.1
cryptography==41.0.7
dataclasses==0.6
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastavro==1.8.2
fastjsonschema==2.19.0
frozenlist==1.4.1
fsspec==2023.12.2
furl==2.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
great-expectations==0.14.13
greenlet==3.0.3
h11==0.14.0
hopsworks==3.4.3
hsfs==3.4.5
hsml==3.4.5
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
ipython==8.18.1
javaobj-py3==0.4.3
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.1
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mock==5.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.9.2
numpy==1.26.2
openai==1.6.1
orderedmultidict==1.0.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyHopsHive==0.6.4.1.dev0
pyhumps==1.6.1
pyjks==20.0.0
PyMySQL==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.29.0
tenacity==8.2.3
termcolor==2.4.0
thrift==0.16.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
twofish==0.3.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
wcwidth==0.2.12
yarl==1.9.4
zipp==3.17.0
